AlbumSearchActivity: Convert Java File to Kotlin File - interim 1b, 1c
1b) Continue "Getting started with Android and Kotlin" by invoking the
"Configure Kotlin in Project" action, then do "Sync Now" to Sync
Project with Gradle Files.

1c) Code Cleanup
The Kotlin version is set to just-released M13, and there are many
changes from step 1a that are now broken or deprecated, so do code
cleanup with IDE menu: Analyze | Code Cleanup on the activity file.

Caveat -- build will still not succeed at this point because of
Butterknife's field injection (which is for Java code and not Kotlin).

Next interim commit: replace Butterknife with Kotlin Android Extensions
and fix issue of non-null type val that is injected and hence not
initialized.