Use Kotlin Android Extensions, Use Late-init - interim commit 2 - 5
2) Follow "Kotlin Android Extensions"
   http://kotlinlang.org/docs/tutorials/android-plugin.html
to use Kotlin Android Extensions plugin to replace ButterKnife's @InjectView.
This works for the TextView with id "empty_view" in the layout
    <TextView
        android:id="@+id/empty_view"
        ...
but does not appear to work for the ListView with android-domain id "list"
    <ListView
        android:id="@android:id/list"
        ...
for some reason, so we initialize that ListView the old fashion way with
findViewById(int) for now.

3) Use new feature in M13 of late-init properties so that we can declare a
non-null var without an initializer, because it will be injected later.

4) Others:
- Update gradle plugin version to latest 1.3.0.
- Reformat code put in by step 1a "Convert Java File to Kotlin File" action.

5) Proguard entries
Add entries to try to fix BUILD FAILED when we do Run | Run 'app'
- extraneous warnings, and a
- :app:packageDebug FAILED  > Unable to compute hash of ... classes.jar.
This effort is complicated by a new Proguard bug exposed by M13 release of
Kotlin which causes following to be also seen before packageDebug action:
- :app:proguardDebug  Exception while processing task
   java.lang.StringIndexOutOfBoundsException: String index out of range: 37

Caveat - a Build | Build Project will succeed, but Run | Run 'app' will fail
because of Proguard at this time.

Next commit - go back to M12 Kotlin release, and work-around lack of lateinit
feature for now until Proguard bug is fixed.