Cache observable of album items from Retrofit api call
Store the cache as static field in Application subclass so it
survives Activity lifecycle.

If cache is null, make the api call and cache the observable
obtained after processing to get bitmaps from the thumbnail urls
and to get the albums' name; then subscribe to the newly stored
observable so as display the album items in the recyclerview.

But if cache is not null, just subscribe to the cached observable
so that we can re-use it without making the api call again.

When Activity is created, say after an orientation change, check
whether cache is null, and if not, subscribe to the cached
observable, so as display the album items.

To be done - clearing the adapter/recyclerview and cache.