Improve behavior when there is a cached observable
In Activity onCreate(), add check that cached observable is not
empty before using it to display the album items. Here, also add
code to hide the textview that has prompt-text which otherwise can
be seen behind the recyclerview when there are gaps in the list of
card views.

Some refactorings:
- rename the adapter class since we now don't use gridview
- encapsulate code to set prompt-textview visibility in a method
- encapsulate code to subscribe to an observable in a method
The encapsulation allows for more DRY code.