Refactor to use safe operator ?. instead of !! operator
Mostly straight replacement, except for calling adapter's
notifyItemInserted(). It needs a non-null argument so
need to 1) replace the "-" operator with minus() function
called with safe operator ?. and 2) use the elvis operator
to ensure a non-null argument is passed in.

Separate out the subscribing and the adding to the
CompositeSubscription steps, for easier to read code.